Fixes the compile-time failure with tiny go
Added an Initialization function to the whole library to avoid trivial compile-time errors. This can also be used in the future for more complicated Initialization logic that shouldn't be done in the global scope and/or is error prone.

For more infomation, see https://github.com/tinygo-org/tinygo/issues/727#issuecomment-554582530